[052_modular_approach], issue #52, Moved 'Fillable' class in a dedicated module
